🐛 fix : pin 좌표 Float 형으로 변경

+++ backend/src/main/java/com/dawn/backend/domain/pin/dto/PinItem.java
@@ -5,8 +5,8 @@ import java.util.List;
 public record PinItem(
 	Long pinId,
 	String pinName,
-	Integer pinX,
-	Integer pinY,
+	Float pinX,
+	Float pinY,
 	List<ImageItem> imageList,
 	Integer imageCount,
 	PinGroupDto pinGroup,
